chore: separated functions in components
